moved some things around
