+++ packages/client/src/styles.less
@@ -49,6 +49,7 @@ mat-sidenav-container {
     background-color: var(--site-accent);
     color: whitesmoke;
     border-right: 1px solid var(--site-borders);
+    border-left: 1px solid var(--site-borders);
 }
 
 div.user-buttons {
